Introduce singleton comparer. Introduce factory methods for inclusive/exclusive validators.
