+++ src/FluentValidation/Validators/ComparableComparer.cs
@@ -19,8 +19,15 @@
 namespace FluentValidation.Validators {
 	using System;
 	using System.Collections.Generic;
+	using System.Drawing;
 
 	internal class ComparableComparer<T> : IComparer<T> where T : IComparable<T> {
+		internal static ComparableComparer<T> Instance { get; }
+
+		static ComparableComparer() {
+			Instance = new ComparableComparer<T>();
+		}
+
 		public int Compare(T x, T y) {
 			return x.CompareTo(y);
 		}
